Fix erratic point addition in Trigger

+++ app/javascript/components/customers/ordersubmission/CompleteOrder.js
@@ -62,7 +62,7 @@ class CompleteOrder extends React.Component {
                                 : 0;
         order["payment_method"] = this.state.paymentMethod;
         order["delivery_fee"] = 3.00
-        order["total_price"] = parseInt(sessionStorage.getItem('amount_due'));
+        order["total_price"] = parseFloat(sessionStorage.getItem('amount_due'));
         order["status"] = "in progress";
         order["foods"] = JSON.parse(sessionStorage.getItem('foods'));
         order["customer_location"] = this.state.address;
